+++ Introspect/ViewExtensions.swift
@@ -6,7 +6,7 @@ import AppKit
 import UIKit
 #endif
 
-@available(iOS 13.0, tvOS 13.0, macOS 15.0, *)
+@available(iOS 13.0, tvOS 13.0, macOS 10.15.0, *)
 extension View {
     public func inject<SomeView>(_ view: SomeView) -> some View where SomeView: View {
         return overlay(view.frame(width: 0, height: 0))
@@ -14,7 +14,7 @@ extension View {
 }
 
 #if canImport(UIKit)
-@available(iOS 13.0, tvOS 13.0, macOS 15.0, *)
+@available(iOS 13.0, tvOS 13.0, macOS 10.15.0, *)
 extension View {
     
     /// Finds a `TargetView` from a `SwiftUI.View`
@@ -72,57 +72,61 @@ extension View {
     
     /// Finds a `UITableView` from a `SwiftUI.List`, or `SwiftUI.List` child.
     public func introspectTableView(customize: @escaping (UITableView) -> ()) -> some View {
-        return introspect(selector: TargetViewSelector.ancestorOrSibling, customize: customize)
+        return introspect(selector: TargetViewSelector.ancestorOrSiblingContaining, customize: customize)
     }
     
     /// Finds a `UIScrollView` from a `SwiftUI.ScrollView`, or `SwiftUI.ScrollView` child.
     public func introspectScrollView(customize: @escaping (UIScrollView) -> ()) -> some View {
-        return introspect(selector: TargetViewSelector.ancestorOrSibling, customize: customize)
+        if #available(iOS 14.0, tvOS 14.0, macOS 11.0, *) {
+            return introspect(selector: TargetViewSelector.ancestorOrSiblingOfType, customize: customize)
+        } else {
+            return introspect(selector: TargetViewSelector.ancestorOrSiblingContaining, customize: customize)
+        }
     }
     
     /// Finds a `UITextField` from a `SwiftUI.TextField`
     public func introspectTextField(customize: @escaping (UITextField) -> ()) -> some View {
-        return introspect(selector: TargetViewSelector.sibling, customize: customize)
+        return introspect(selector: TargetViewSelector.siblingContaining, customize: customize)
     }
 
     /// Finds a `UITextView` from a `SwiftUI.TextEditor`
     public func introspectTextEditor(customize: @escaping (UITextView) -> ()) -> some View {
-        return introspect(selector: TargetViewSelector.sibling, customize: customize)
+        return introspect(selector: TargetViewSelector.siblingContaining, customize: customize)
     }
     
     /// Finds a `UISwitch` from a `SwiftUI.Toggle`
     @available(tvOS, unavailable)
     public func introspectSwitch(customize: @escaping (UISwitch) -> ()) -> some View {
-        return introspect(selector: TargetViewSelector.sibling, customize: customize)
+        return introspect(selector: TargetViewSelector.siblingContaining, customize: customize)
     }
     
     /// Finds a `UISlider` from a `SwiftUI.Slider`
     @available(tvOS, unavailable)
     public func introspectSlider(customize: @escaping (UISlider) -> ()) -> some View {
-        return introspect(selector: TargetViewSelector.sibling, customize: customize)
+        return introspect(selector: TargetViewSelector.siblingContaining, customize: customize)
     }
     
     /// Finds a `UIStepper` from a `SwiftUI.Stepper`
     @available(tvOS, unavailable)
     public func introspectStepper(customize: @escaping (UIStepper) -> ()) -> some View {
-        return introspect(selector: TargetViewSelector.sibling, customize: customize)
+        return introspect(selector: TargetViewSelector.siblingContaining, customize: customize)
     }
     
     /// Finds a `UIDatePicker` from a `SwiftUI.DatePicker`
     @available(tvOS, unavailable)
     public func introspectDatePicker(customize: @escaping (UIDatePicker) -> ()) -> some View {
-        return introspect(selector: TargetViewSelector.sibling, customize: customize)
+        return introspect(selector: TargetViewSelector.siblingContaining, customize: customize)
     }
     
     /// Finds a `UISegmentedControl` from a `SwiftUI.Picker` with style `SegmentedPickerStyle`
     public func introspectSegmentedControl(customize: @escaping (UISegmentedControl) -> ()) -> some View {
-        return introspect(selector: TargetViewSelector.sibling, customize: customize)
+        return introspect(selector: TargetViewSelector.siblingContaining, customize: customize)
     }
 }
 #endif
 
 #if canImport(AppKit) && !targetEnvironment(macCatalyst)
-@available(macOS 15.0, *)
+@available(macOS 10.15.0, *)
 extension View {
     
     /// Finds a `TargetView` from a `SwiftUI.View`
@@ -138,37 +142,37 @@ extension View {
     
     /// Finds a `NSTableView` from a `SwiftUI.List`, or `SwiftUI.List` child.
     public func introspectTableView(customize: @escaping (NSTableView) -> ()) -> some View {
-        return introspect(selector: TargetViewSelector.ancestorOrSibling, customize: customize)
+        return introspect(selector: TargetViewSelector.ancestorOrSiblingContaining, customize: customize)
     }
     
     /// Finds a `NSScrollView` from a `SwiftUI.ScrollView`, or `SwiftUI.ScrollView` child.
     public func introspectScrollView(customize: @escaping (NSScrollView) -> ()) -> some View {
-        return introspect(selector: TargetViewSelector.ancestorOrSibling, customize: customize)
+        return introspect(selector: TargetViewSelector.ancestorOrSiblingContaining, customize: customize)
     }
     
     /// Finds a `NSTextField` from a `SwiftUI.TextField`
     public func introspectTextField(customize: @escaping (NSTextField) -> ()) -> some View {
-        return introspect(selector: TargetViewSelector.sibling, customize: customize)
+        return introspect(selector: TargetViewSelector.siblingContaining, customize: customize)
     }
     
     /// Finds a `NSSlider` from a `SwiftUI.Slider`
     public func introspectSlider(customize: @escaping (NSSlider) -> ()) -> some View {
-        return introspect(selector: TargetViewSelector.sibling, customize: customize)
+        return introspect(selector: TargetViewSelector.siblingContaining, customize: customize)
     }
     
     /// Finds a `NSStepper` from a `SwiftUI.Stepper`
     public func introspectStepper(customize: @escaping (NSStepper) -> ()) -> some View {
-        return introspect(selector: TargetViewSelector.sibling, customize: customize)
+        return introspect(selector: TargetViewSelector.siblingContaining, customize: customize)
     }
     
     /// Finds a `NSDatePicker` from a `SwiftUI.DatePicker`
     public func introspectDatePicker(customize: @escaping (NSDatePicker) -> ()) -> some View {
-        return introspect(selector: TargetViewSelector.sibling, customize: customize)
+        return introspect(selector: TargetViewSelector.siblingContaining, customize: customize)
     }
     
     /// Finds a `NSSegmentedControl` from a `SwiftUI.Picker` with style `SegmentedPickerStyle`
     public func introspectSegmentedControl(customize: @escaping (NSSegmentedControl) -> ()) -> some View {
-        return introspect(selector: TargetViewSelector.sibling, customize: customize)
+        return introspect(selector: TargetViewSelector.siblingContaining, customize: customize)
     }
 }
 #endif